+++ src/converterRule.ts
@@ -1,6 +1,7 @@
 import Parser from '../index';
 import {Token} from './index';
 import {AtomToken} from './atom';
+import type {AST} from '../base';
 import type {ConverterToken, ConverterFlagsToken} from '../internal';
 
 /* NOT FOR BROWSER */
@@ -42,12 +43,12 @@ export abstract class ConverterRuleToken extends Token {
 		return 'converter-rule';
 	}
 
-	/* NOT FOR BROWSER */
-
 	/** language variant / 语言变体 */
 	get variant(): string {
 		return this.childNodes[this.length - 2]?.text().trim() ?? '';
 	}
+
+	/* NOT FOR BROWSER */
 
 	set variant(variant) {
 		this.setVariant(variant);
@@ -157,6 +158,13 @@ export abstract class ConverterRuleToken extends Token {
 			return `<span class="wpb-converter-rule">${from.print()}=>${variant.print()}:${to.print()}</span>`;
 		}
 		return super.print({sep: ':'});
+	}
+
+	/** @private */
+	override json(_?: string, start = this.getAbsoluteIndex()): AST {
+		const json = super.json(undefined, start);
+		json['variant'] = this.variant;
+		return json;
 	}
 
 	/* NOT FOR BROWSER */