+++ src/link/category.ts
@@ -1,10 +1,11 @@
+import {decodeHtml} from '../../util/string';
 import {LinkBaseToken} from './base';
 import type {Title} from '../../lib/title';
+import type {AST} from '../../base';
 import type {Token, AtomToken} from '../../internal';
 
 /* NOT FOR BROWSER */
 
-import {decodeHtml} from '../../util/string';
 import {classes} from '../../util/constants';
 
 /* NOT FOR BROWSER END */
@@ -30,17 +31,31 @@ export abstract class CategoryToken extends LinkBaseToken {
 		return 'category';
 	}
 
-	/* NOT FOR BROWSER */
-
 	/** sort key / 分类排序关键字 */
 	get sortkey(): string | undefined {
 		const {childNodes: [, child]} = this;
 		return child && decodeHtml(child.text());
 	}
 
+	/* NOT FOR BROWSER */
+
 	set sortkey(text) {
 		this.setSortkey(text);
 	}
+
+	/* NOT FOR BROWSER END */
+
+	/** @private */
+	override json(_?: string, start = this.getAbsoluteIndex()): AST {
+		const json = super.json(undefined, start),
+			{sortkey} = this;
+		if (sortkey) {
+			json['sortkey'] = sortkey;
+		}
+		return json;
+	}
+
+	/* NOT FOR BROWSER */
 
 	/**
 	 * Set the sort key